Spawning workers outside liftAuth
