+++ wallet/src/Ergvein/Wallet/Monad/Auth.hs
@@ -1,4 +1,6 @@
+{-# OPTIONS_GHC -Wall #-}
 {-# LANGUAGE UndecidableInstances #-}
+
 module Ergvein.Wallet.Monad.Auth(
     liftAuth
   , liftUnauthed
@@ -7,23 +9,20 @@ module Ergvein.Wallet.Monad.Auth(
 import Control.Concurrent
 import Control.Concurrent.Chan (Chan)
 import Control.Concurrent.STM.TChan
-import Control.Concurrent.STM.TVar
 import Control.Lens
 import Control.Monad.STM
 import Control.Monad.Reader
 import Data.Map.Strict (Map)
 import Data.Fixed
 import Data.Text as T
-import Data.Time (getCurrentTime, diffUTCTime, NominalDiffTime)
+import Data.Time (NominalDiffTime)
 import Network.Socket
 import Reflex
 import Reflex.Dom
 import Reflex.Dom.Retractable
 import Reflex.ExternalRef
-import System.Directory
 
 import Ergvein.Node.Constants
-import Ergvein.Node.Resolve
 import Ergvein.Types.AuthInfo
 import Ergvein.Types.Currency
 import Ergvein.Types.Fees
@@ -33,25 +32,16 @@ import Ergvein.Types.Storage
 import Ergvein.Types.Transaction (BlockHeight)
 import Ergvein.Wallet.Language
 import Ergvein.Wallet.Log.Types
-import Ergvein.Wallet.Monad.Async
 import Ergvein.Wallet.Monad.Client
 import Ergvein.Wallet.Monad.Front
 import Ergvein.Wallet.Monad.Storage
 import Ergvein.Wallet.Native
 import Ergvein.Wallet.Node
-import Ergvein.Wallet.Platform
-import Ergvein.Wallet.Scan
 import Ergvein.Wallet.Settings (Settings(..))
 import Ergvein.Wallet.Status.Types
 import Ergvein.Wallet.Storage.Util
 import Ergvein.Wallet.Version
-import Ergvein.Wallet.Worker.Fees
-import Ergvein.Wallet.Worker.Height
-import Ergvein.Wallet.Worker.Node
-import Ergvein.Wallet.Worker.PubKeysGenerator
-import Ergvein.Wallet.Worker.Rates
 
-import qualified Data.List as L
 import qualified Data.Map.Strict as M
 import qualified Data.Set as S
 import qualified Data.Vector as V
@@ -261,53 +251,6 @@ instance (MonadBaseConstr t m, HasStoreDir m) => MonadStorage t (ErgveinM t m) w
   getStoreChan = asks env'storeChan
   {-# INLINE getStoreChan #-}
 
--- | Minimum time between two writes of storage to disk
-storeTimeBetweenWrites :: NominalDiffTime
-storeTimeBetweenWrites = 20
-
--- | Thread that writes down updates of wallet storages and checks that write down doesn't occur too frequent.
-walletStoreThread :: PlatformNatives => Text -> MVar () -> TChan (Text, AuthInfo) -> IO ()
-walletStoreThread storeDir mutex updChan = void $ forkOnOther $ do
-  timeRef <- newTVarIO =<< getCurrentTime
-  lastUpdTimeRef <- newTVarIO =<< getCurrentTime
-  lastStoreRef <- newTVarIO Nothing
-  -- Thread that updates reference with time to compare it with value in lastUpdTimeRef in getTimedWrite
-  void $ forkIO $ fix $ \next -> do
-    threadDelay $ ceiling storeTimeBetweenWrites
-    currTime <- getCurrentTime
-    atomically $ writeTVar timeRef currTime
-    next
-  -- Thread that reads from chan and stores last storage to reference which next thread will check and validate
-  -- against timeout.
-  void $ forkIO $ fix $ \next -> do
-    atomically $ do
-      val <- readTChan updChan
-      writeTVar lastStoreRef $ Just val
-    next
-  -- If we have awaiting write to disk and time passed > timeout we return the value unless retry
-  let getTimedWrite = do
-        mval <- readTVar lastStoreRef
-        case mval of
-          Nothing -> retry
-          Just val -> do
-            currTime <- readTVar timeRef
-            updTime <- readTVar lastUpdTimeRef
-            when (diffUTCTime currTime updTime < storeTimeBetweenWrites) retry
-            writeTVar lastUpdTimeRef currTime
-            writeTVar lastStoreRef Nothing
-            pure val
-  -- Thread that indefinetely queries if we need to write down new state
-  fix $ \next -> do
-    (caller, authInfo) <- atomically getTimedWrite
-    storeWalletIO caller storeDir mutex authInfo
-    next
-
-storeWalletIO :: PlatformNatives => Text -> Text -> MVar () -> AuthInfo -> IO ()
-storeWalletIO caller storeDir mutex ai = do
-  let storage = _authInfo'storage ai
-  let eciesPubKey = _authInfo'eciesPubKey ai
-  withMVar mutex $ const $ flip runReaderT storeDir $ saveStorageToFile caller eciesPubKey storage
-
 -- | Execute action under authorized context or return the given value as result
 -- if user is not authorized. Each time the login info changes and authInfo'isUpdate flag is set to 'False'
 -- (user logs out or logs in) the widget is updated.
@@ -401,18 +344,6 @@ liftAuth ma0 ma = mdo
               , env'indexReqFire = iReqFire
               , env'activateIndexEF = indexEF
               }
-
-        flip runReaderT env $ do -- Workers and other routines go here
-          liftIO $ walletStoreThread storeDir storeMutex storeChan
-          when isAndroid (deleteTmpFiles storeDir)
-          -- initFiltersHeights filtersHeights
-          scanner
-          btcNodeController
-          ratesWorker
-          heightAsking
-          feesWorker
-          pubKeysGenerator
-          pure ()
         runReaderT (wrapped "liftAuth" ma) env
   let
     ma0' = maybe ma0 runAuthed mauth0
@@ -437,10 +368,3 @@ wrapped caller ma = do
   void . updateActiveCurs $ fmap (\cl -> const (S.fromList cl)) $ ac <$ buildE
   ma
   where clr = caller <> ":" <> "wrapped"
-
--- Deletes files created with 'atomicWriteFile' from specified directiry
-deleteTmpFiles :: MonadIO m => Text -> m ()
-deleteTmpFiles dir = liftIO $ do
-  entries <- listDirectory $ T.unpack dir
-  traverse_ removeFile $ L.filter isTmpFile entries
-  where isTmpFile filePath = "atomic" `L.isPrefixOf` filePath && ".write" `L.isSuffixOf` filePath
